Replaced RTL and BLIF generator with a backend generator
