+++ include/dynamatic/Support/BackendGenerator.h
@@ -1,0 +1,103 @@
+//===- BackendGenerator.h - RTL and BLIF backend generator ------*- C++ -*-===//
+//
+// Dynamatic is under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+//
+// Generates RTL (Verilog) or BLIF output for a given Handshake operation.
+// Parameters are extracted from the operation via RTLAttrInterface and matched
+// against a JSON config to locate and execute the appropriate generator
+// command.
+//
+// For the Verilog backend the generator command produces .v files directly.
+// For the BLIF backend the generator first produces Verilog, then synthesizes
+// it to BLIF via Yosys and optimizes the result with ABC.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef DYNAMATIC_SUPPORT_BACKENDGENERATOR_H
+#define DYNAMATIC_SUPPORT_BACKENDGENERATOR_H
+
+#include <filesystem>
+#include <string>
+#include <variant>
+#include <vector>
+
+namespace mlir {
+class Operation;
+} // namespace mlir
+
+namespace dynamatic {
+
+class BackendGenerator {
+public:
+  enum class Backend { Verilog, BLIF };
+
+  /// Parameters for Verilog generation: run the Verilog generator command and
+  /// collect the produced .v files.
+  struct VerilogParams {
+    std::string rtlConfigPath; /// Absolute path to the JSON RTL config.
+    std::string dynamaticRoot; /// Project root for $DYNAMATIC substitution.
+    std::string outputBaseDir; /// Base directory for per-module output.
+  };
+
+  /// Parameters for BLIF generation: call Verilog generator, then
+  /// synthesize to BLIF via Yosys and optimize with ABC.
+  struct BLIFParams {
+    std::string rtlConfigPath;   /// Absolute path to the JSON RTL config.
+    std::string dynamaticRoot;   /// Project root for $DYNAMATIC substitution.
+    std::string outputBaseDir;   /// Base directory for per-module output.
+    std::string yosysExecutable; /// Path to the Yosys binary.
+    std::string abcExecutable;   /// Path to the ABC binary.
+  };
+
+  using Params = std::variant<VerilogParams, BLIFParams>;
+
+  BackendGenerator(Backend backend, Params params);
+
+  /// Run Verilog generation for op and, for the BLIF backend, Yosys+ABC
+  /// synthesis. On success, results are available via the getters below.
+  bool generate(mlir::Operation *op);
+
+  const std::string &getModuleName() const { return moduleName; }
+  /// Verilog backend: returns the generated .v files.
+  /// BLIF backend: returns a single-element vector with the .blif file path.
+  const std::vector<std::string> &getOutputFiles() const { return outputFiles; }
+
+private:
+  Backend backend; /// The target backend to generate for (Verilog or BLIF).
+  Params params;   /// Parameters for generation, depending on the backend.
+
+  std::string moduleName; /// The name of the generated module, derived from the
+                          /// operation name.
+  std::vector<std::string>
+      outputFiles; /// Paths of the generated files (.v for Verilog backend,
+                   /// .blif for BLIF backend).
+
+  /// Get the directory where the backend generator will place output files for
+  /// operation op.
+  std::filesystem::path getOutputDirForOp(mlir::Operation *op,
+                                          const std::string &baseDir);
+
+  /// Verilog backend: generate RTL and store .v files in outputFiles.
+  bool generateVerilog(mlir::Operation *op);
+
+  /// BLIF backend: generate RTL then run Yosys+ABC, store .blif in outputFiles.
+  bool generateBLIF(mlir::Operation *op);
+
+  /// Run Yosys to synthesize the generated Verilog files into BLIF.
+  bool runYosys(const std::string &yosysExecutable,
+                const std::string &topModule, const std::string &outputDir,
+                const std::vector<std::string> &verilogFiles,
+                const std::string &outputName) const;
+  /// Run ABC to optimize the BLIF file generated by Yosys.
+  bool runAbc(const std::string &abcExecutable, const std::string &inputFile,
+              const std::string &outputDir,
+              const std::string &outputName) const;
+};
+
+} // namespace dynamatic
+
+#endif // DYNAMATIC_SUPPORT_BACKENDGENERATOR_H
